Correct field type

+++ app/src/main/java/com/sergioborne/themoviedb/common/data/TvShow.kt
@@ -4,7 +4,7 @@ data class TvShow(
     val id: Int,
     val name: String,
     val poster_path: String,
-    val vote_average: Long
+    val vote_average: Double
 )
